Refactor: extract function-parameter-naming module from print.ts
