+++ src/format/src/printer/print.ts
@@ -41,7 +41,18 @@ import {
     UNDEFINED_TYPE
 } from "./constants.js";
 import { getEnumNameAlignmentPadding, prepareEnumMembersForPrinting } from "./enum-alignment.js";
-import { findAncestorNode, safeGetParentNode } from "./path-utils.js";
+import {
+    filterKeptDeclarators,
+    filterMisattachedFunctionDocComments,
+    findEnclosingFunctionDeclaration,
+    findEnclosingFunctionNode,
+    getPreferredFunctionParameterName,
+    joinDeclaratorPartsWithCommas,
+    resolveArgumentAliasInitializerDoc,
+    resolvePreferredParameterName,
+    shouldSynthesizeUndefinedDefaultForIdentifier
+} from "./function-parameter-naming.js";
+import { safeGetParentNode } from "./path-utils.js";
 import {
     breakParent,
     concat,
@@ -91,8 +102,7 @@ import {
     isLogicalComparisonClause,
     isNumericComputationNode,
     isSimpleCallArgument,
-    isSyntheticParenFlatteningEnabled,
-    isValidIdentifierName
+    isSyntheticParenFlatteningEnabled
 } from "./type-guards.js";
 
 // TODO: Use Core.* directly instead of destructuring the Core namespace across
@@ -2494,341 +2504,6 @@ function findNextTerminalCharacter(
     return null;
 }
 
-function getPreferredFunctionParameterName(path, node, options) {
-    const context = findFunctionParameterContext(path);
-    if (context) {
-        const { functionNode, paramIndex } = context;
-        if (!functionNode || !Number.isInteger(paramIndex) || paramIndex < 0) {
-            return null;
-        }
-
-        const params = getFunctionParams(functionNode);
-        if (paramIndex >= params.length) {
-            return null;
-        }
-
-        const identifier = Core.getIdentifierFromParameterNode(paramIndex);
-        const currentName =
-            (identifier && typeof identifier.name === STRING_TYPE ? identifier.name : null) ??
-            (node && typeof node.name === STRING_TYPE ? node.name : null);
-
-        const preferredName = resolvePreferredParameterName(functionNode, paramIndex, currentName, options);
-
-        if (Core.isNonEmptyString(preferredName)) {
-            return preferredName;
-        }
-
-        return null;
-    }
-
-    if (!node || typeof node.name !== STRING_TYPE) {
-        return null;
-    }
-
-    const argumentIndex = Core.getArgumentIndexFromIdentifier(node.name);
-    if (!Number.isInteger(argumentIndex) || argumentIndex < 0) {
-        return null;
-    }
-
-    const functionNode = findEnclosingFunctionNode(path);
-    if (!functionNode) {
-        return null;
-    }
-
-    const preferredName = resolvePreferredParameterName(functionNode, argumentIndex, node.name, options);
-
-    if (Core.isNonEmptyString(preferredName)) {
-        return preferredName;
-    }
-
-    const params = getFunctionParams(functionNode);
-    if (argumentIndex >= params.length) {
-        return null;
-    }
-
-    const identifier = Core.getIdentifierFromParameterNode(params[argumentIndex]);
-    if (!identifier || typeof identifier.name !== STRING_TYPE) {
-        return null;
-    }
-
-    const normalizedIdentifier = normalizePreferredParameterName(identifier.name);
-    if (normalizedIdentifier && normalizedIdentifier !== node.name && isValidIdentifierName(normalizedIdentifier)) {
-        return normalizedIdentifier;
-    }
-
-    return null;
-}
-
-function resolvePreferredParameterName(functionNode, paramIndex, currentName, options) {
-    if (!functionNode || !Number.isInteger(paramIndex) || paramIndex < 0) {
-        return null;
-    }
-
-    const params = getFunctionParams(functionNode);
-    if (paramIndex >= params.length) {
-        return null;
-    }
-
-    const functionTagName = getFunctionTagParamName(functionNode, paramIndex, options);
-    const hasRenamableCurrentName =
-        typeof currentName === STRING_TYPE && Core.getArgumentIndexFromIdentifier(currentName) !== null;
-
-    if (!hasRenamableCurrentName) {
-        return null;
-    }
-
-    const preferredSource = resolvePreferredParameterSource(
-        functionNode,
-        paramIndex,
-        currentName,
-        options,
-        functionTagName
-    );
-
-    const normalizedName = normalizePreferredParameterName(preferredSource);
-    if (!normalizedName || normalizedName === currentName) {
-        return null;
-    }
-
-    return isValidIdentifierName(normalizedName) ? normalizedName : null;
-}
-
-function resolvePreferredParameterSource(functionNode, paramIndex, currentName, options, functionTagName) {
-    if (Core.isNonEmptyString(functionTagName)) {
-        return functionTagName;
-    }
-
-    const docPreferences = Core.preferredParamDocNamesByNode.get(functionNode);
-    if (docPreferences?.has(paramIndex)) {
-        return docPreferences.get(paramIndex) ?? null;
-    }
-
-    const implicitEntries = Core.collectImplicitArgumentDocNames(functionNode, options);
-    if (!Array.isArray(implicitEntries)) {
-        return null;
-    }
-
-    const implicitEntry = implicitEntries.find((entry) => entry && entry.index === paramIndex);
-    if (!implicitEntry) {
-        return null;
-    }
-
-    if (implicitEntry.canonical) {
-        return implicitEntry.name || implicitEntry.canonical;
-    }
-
-    if (implicitEntry.name && implicitEntry.name !== currentName) {
-        return implicitEntry.name;
-    }
-
-    return null;
-}
-
-function getFunctionTagParamName(functionNode, paramIndex, options) {
-    if (!functionNode || !Number.isInteger(paramIndex) || paramIndex < 0) {
-        return null;
-    }
-
-    const orderedParamNames = Array.isArray(functionNode?._functionTagParamNames)
-        ? functionNode._functionTagParamNames
-        : null;
-    if (
-        orderedParamNames &&
-        paramIndex < orderedParamNames.length &&
-        Core.isNonEmptyString(orderedParamNames[paramIndex])
-    ) {
-        return orderedParamNames[paramIndex];
-    }
-
-    const docComments = Array.isArray(functionNode.docComments)
-        ? functionNode.docComments
-        : Array.isArray(functionNode.comments)
-          ? functionNode.comments
-          : null;
-    if (!Core.isNonEmptyArray(docComments)) {
-        return null;
-    }
-
-    const lineCommentOptions = Core.resolveLineCommentOptions(options);
-    const formattingOptions = {
-        ...lineCommentOptions,
-        originalText: options?.originalText
-    };
-
-    for (const comment of docComments) {
-        const formatted = Core.formatLineComment(comment, formattingOptions);
-        const rawValue = formatted ?? (typeof comment?.value === "string" ? comment.value : null);
-        if (!Core.isNonEmptyString(rawValue)) {
-            continue;
-        }
-
-        const params = Core.extractFunctionTagParams(rawValue);
-        if (params.length === 0) {
-            continue;
-        }
-
-        return paramIndex < params.length ? params[paramIndex] : null;
-    }
-
-    const docParamFromOriginalText = getFunctionTagParamFromOriginalText(
-        options?.originalText,
-        functionNode,
-        paramIndex
-    );
-    if (docParamFromOriginalText !== null) {
-        return docParamFromOriginalText;
-    }
-
-    return null;
-}
-
-function getFunctionTagParamFromOriginalText(
-    originalText: string | null | undefined,
-    functionNode: unknown,
-    paramIndex: number
-): string | null {
-    if (typeof originalText !== STRING_TYPE) {
-        return null;
-    }
-
-    const functionStart = Core.getNodeStartIndex(functionNode);
-    if (typeof functionStart !== NUMBER_TYPE) {
-        return null;
-    }
-
-    const prefix = originalText.slice(0, functionStart);
-    const lastDocIndex = prefix.lastIndexOf("///");
-    if (lastDocIndex === -1) {
-        return null;
-    }
-
-    const docBlock = prefix.slice(lastDocIndex);
-    const lines = Core.splitLines(docBlock);
-    for (const docLine of lines) {
-        const params = Core.extractFunctionTagParams(docLine);
-        if (params.length > 0) {
-            return paramIndex < params.length ? params[paramIndex] : null;
-        }
-    }
-
-    return null;
-}
-
-function findEnclosingFunctionNode(path) {
-    return findAncestorNode(path, (node) => Core.isFunctionLikeDeclaration(node));
-}
-
-function findFunctionParameterContext(path) {
-    if (!path || typeof path.getParentNode !== "function") {
-        return null;
-    }
-
-    let candidate = path.getValue();
-    for (let depth = 0; ; depth += 1) {
-        const parent = safeGetParentNode(path, depth);
-        if (!parent) {
-            break;
-        }
-
-        if (parent.type === "DefaultParameter") {
-            candidate = parent;
-            continue;
-        }
-
-        if (parent.type === "FunctionDeclaration" || parent.type === "ConstructorDeclaration") {
-            const params = Core.toMutableArray(parent.params);
-            const index = params.indexOf(candidate);
-            if (index !== -1) {
-                return { functionNode: parent, paramIndex: index };
-            }
-        }
-
-        candidate = parent;
-    }
-
-    return null;
-}
-
-function shouldOmitParameterAlias(declarator, functionNode, options) {
-    if (
-        !declarator ||
-        declarator.type !== "VariableDeclarator" ||
-        !declarator.id ||
-        declarator.id.type !== "Identifier" ||
-        !declarator.init ||
-        declarator.init.type !== "Identifier"
-    ) {
-        return false;
-    }
-
-    const aliasName = declarator.id.name;
-
-    const normalizedAliasName = normalizePreferredParameterName(aliasName);
-    const normalizedInitName = normalizePreferredParameterName(declarator.init.name);
-
-    if (normalizedAliasName && normalizedInitName && normalizedAliasName === normalizedInitName) {
-        return true;
-    }
-
-    const argumentIndex = Core.getArgumentIndexFromIdentifier(declarator.init.name);
-
-    let paramIndex = argumentIndex;
-    if (argumentIndex === null && functionNode) {
-        const params = getFunctionParams(functionNode);
-        for (const [i, param] of params.entries()) {
-            const paramId = Core.getIdentifierFromParameterNode(param);
-            if (paramId && paramId.name === declarator.init.name) {
-                paramIndex = i;
-                break;
-            }
-        }
-        if (paramIndex === null) {
-            return false;
-        }
-    } else if (argumentIndex === null) {
-        return false;
-    }
-
-    const preferredName = resolvePreferredParameterName(functionNode, paramIndex, declarator.init.name, options);
-
-    const normalizedAlias = normalizePreferredParameterName(aliasName);
-    if (!normalizedAlias) {
-        return false;
-    }
-
-    if (!functionNode) {
-        return false;
-    }
-
-    const params = getFunctionParams(functionNode);
-    if (paramIndex < 0 || paramIndex >= params.length) {
-        return false;
-    }
-
-    const identifier = Core.getIdentifierFromParameterNode(params[paramIndex]);
-    if (!identifier || typeof identifier.name !== STRING_TYPE) {
-        return false;
-    }
-
-    const normalizedParamName = normalizePreferredParameterName(identifier.name);
-
-    if (
-        typeof normalizedParamName === STRING_TYPE &&
-        normalizedParamName.length > 0 &&
-        normalizedParamName === normalizedAlias
-    ) {
-        return true;
-    }
-
-    const normalizedPreferred = preferredName ? normalizePreferredParameterName(preferredName) : null;
-
-    if (normalizedPreferred && normalizedPreferred === normalizedAlias) {
-        return true;
-    }
-
-    return false;
-}
-
 function printGlobalVarStatementAsKeyword(node, path, print, options) {
     const decls =
         node.declarations.length > 1
@@ -2842,126 +2517,6 @@ function printGlobalVarStatementAsKeyword(node, path, print, options) {
 
     return concat([keyword, " ", decls]);
 }
-
-/**
- * Filters variable declarators based on parameter alias omission rules.
- *
- * Removes declarators that are redundant parameter aliases when formatting optimization
- * is enabled. The filtering logic delegates to `shouldOmitParameterAlias` for each declarator.
- *
- * @param {ReadonlyArray<any>} declarators - Array of declarator nodes to filter
- * @param {any} functionNode - The enclosing function node (if any)
- * @param {any} options - Prettier options
- * @returns {Array<any>} Filtered array of declarators to keep
- */
-function filterKeptDeclarators(declarators, functionNode, options) {
-    return declarators.filter((declarator) => {
-        const omit = shouldOmitParameterAlias(declarator, functionNode, options);
-        return !omit;
-    });
-}
-
-/**
- * Filters out misattached function doc-comments from a declarator's comments array.
- *
- * Mutates the declarator in place by filtering its comments array and marking
- * filtered comments as printed. If all comments are filtered, deletes the comments property.
- *
- * This workaround addresses a parser issue where JSDoc function comments (@function, @func)
- * are incorrectly attached to variable declarators instead of their intended function targets.
- *
- * @param {any} declarator - The variable declarator node to process
- */
-function filterMisattachedFunctionDocComments(declarator) {
-    if (!declarator.comments) {
-        return;
-    }
-
-    declarator.comments = declarator.comments.filter((comment) => {
-        const isFunctionComment = comment.value.includes("@function") || comment.value.includes("@func");
-
-        if (isFunctionComment) {
-            comment.printed = true;
-            return false;
-        }
-
-        return true;
-    });
-
-    if (declarator.comments.length === 0) {
-        delete declarator.comments;
-    }
-}
-
-/**
- * Joins an array of declarator doc fragments with comma separators.
- *
- * Inserts ", " between each pair of elements to produce a comma-separated list
- * suitable for variable declarations.
- *
- * @param {ReadonlyArray<any>} parts - Array of doc fragments to join
- * @returns {Array<any>} Flat array with commas inserted between parts
- */
-function joinDeclaratorPartsWithCommas(parts) {
-    const joined = [];
-    const count = parts.length;
-
-    for (let i = 0; i < count; i += 1) {
-        joined.push(parts[i]);
-
-        if (i < count - 1) {
-            joined.push(", ");
-        }
-    }
-
-    return joined;
-}
-
-function findEnclosingFunctionDeclaration(path) {
-    return findAncestorNode(path, (node) => node.type === "FunctionDeclaration");
-}
-
-function shouldSynthesizeUndefinedDefaultForIdentifier(path, node) {
-    if (!node || Core.synthesizedUndefinedDefaultParameters.has(node)) {
-        return false;
-    }
-
-    if (!path || typeof path.getParentNode !== "function") {
-        return false;
-    }
-
-    const parent = safeGetParentNode(path);
-    if (!parent || parent.type !== "FunctionDeclaration") {
-        return false;
-    }
-
-    const params = getFunctionParams(parent);
-    return params.includes(node);
-}
-
-function normalizePreferredParameterName(name) {
-    if (typeof name !== STRING_TYPE || name.length === 0) {
-        return null;
-    }
-
-    const canonical = Core.getCanonicalParamNameFromText(name);
-    if (canonical && canonical.length > 0) {
-        return canonical;
-    }
-
-    const normalizedValue = Core.normalizeDocMetadataName(name);
-    if (typeof normalizedValue !== STRING_TYPE) {
-        return null;
-    }
-
-    const normalized = (normalizedValue as string).trim();
-    return normalized.length === 0 ? null : normalized;
-}
-
-// Collects index/reference bookkeeping for implicit `arguments[index]` usages
-// within a function. The traversal tracks alias declarations, direct
-// references, and the set of indices that require doc entries so the caller
-// can format them without dipping into low-level mutation logic.
 
 function getSourceTextForNode(node, options) {
     if (!node) {
@@ -4102,111 +3657,6 @@ function shouldPreserveClauseBlockAdjacency(clauseNode, bodyNode) {
 
 function printSimpleDeclaration(leftDoc, rightDoc) {
     return rightDoc ? [leftDoc, " = ", rightDoc] : leftDoc;
-}
-
-function resolveArgumentAliasInitializerDoc(path) {
-    const node = path.getValue();
-    if (!node || node.type !== "VariableDeclarator") {
-        return null;
-    }
-
-    const initializer = node.init;
-    if (!initializer || initializer.type !== "Identifier") {
-        return null;
-    }
-
-    const match = Core.GML_ARGUMENT_IDENTIFIER_PATTERN.exec(initializer.name ?? "");
-    if (!match) {
-        return null;
-    }
-
-    const aliasIdentifier = node.id;
-    if (!aliasIdentifier || aliasIdentifier.type !== "Identifier") {
-        return null;
-    }
-
-    const aliasName = aliasIdentifier.name;
-    if (!Core.isNonEmptyString(aliasName)) {
-        return null;
-    }
-
-    const argumentIndex = Number.parseInt(match[1]);
-    if (!Number.isInteger(argumentIndex) || argumentIndex < 0) {
-        return null;
-    }
-
-    const functionNode = findEnclosingFunctionForPath(path);
-    if (!functionNode) {
-        return null;
-    }
-
-    const docPreferences = Core.preferredParamDocNamesByNode.get(functionNode);
-    let parameterName = null;
-
-    if (docPreferences && docPreferences.has(argumentIndex)) {
-        const preferred = docPreferences.get(argumentIndex);
-        if (Core.isNonEmptyString(preferred)) {
-            parameterName = preferred;
-        }
-    }
-
-    if (!parameterName) {
-        parameterName = getFunctionParameterNameByIndex(functionNode, argumentIndex);
-    }
-
-    if (!parameterName || parameterName === aliasName || parameterName === initializer.name) {
-        return null;
-    }
-
-    return parameterName;
-}
-
-function findEnclosingFunctionForPath(path) {
-    return findAncestorNode(path, (node) => Core.isFunctionLikeNode(node));
-}
-
-function getFunctionParameterNameByIndex(functionNode, index) {
-    if (!functionNode || typeof functionNode !== OBJECT_TYPE) {
-        return null;
-    }
-
-    const params = getFunctionParams(functionNode);
-
-    if (!Number.isInteger(index) || index < 0 || index >= params.length) {
-        return null;
-    }
-
-    const param = params[index];
-    if (!param || typeof param !== OBJECT_TYPE) {
-        return null;
-    }
-
-    if (param.type === "Identifier" && typeof param.name === STRING_TYPE) {
-        return param.name;
-    }
-
-    if (
-        param.type === "DefaultParameter" &&
-        param.left?.type === "Identifier" &&
-        typeof param.left.name === STRING_TYPE
-    ) {
-        return param.left.name;
-    }
-
-    return null;
-}
-
-function getFunctionParams(functionNode) {
-    if (!functionNode || typeof functionNode !== OBJECT_TYPE) {
-        return [];
-    }
-
-    const { params } = functionNode;
-    if (!Array.isArray(params)) {
-        return [];
-    }
-
-    return params;
 }
 
 // prints empty parens with dangling comments
